Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -3,7 +3,7 @@
 
 
 int main() {
-    float a,b,c;
+   int a,b,c;
     float average;
     scanf("%d %d %d",&a,&b,&c);
     average=(a+b+c)/3;
